requirements.txt
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
bcrypt==4.3.0
Brotli==1.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
cssselect2==0.8.0
fastapi==0.115.13
fastapi-utils==0.8.0
fonttools==4.58.4
greenlet==3.2.3
h11==0.16.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mypy_extensions==1.1.0
passlib==1.7.4
pillow==11.2.1
psutil==5.9.8
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydyf==0.11.0
PyJWT==2.10.1
PyMySQL==1.1.1
pyphen==0.17.2
python-multipart==0.0.20
requests==2.32.4
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.46.2
tinycss2==1.4.0
tinyhtml5==2.0.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.34.3
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1